Faster Sorting for Large Trees
For graphs with text labelled segments, having the arrows flow from left to right makes more readable. The random() function was replaced with a random near diagonal. This assumes that the data node parrents are created before node children. This also tends to create a more stable and quicker sort.